split out scrollbar, add scrollbar to app grid
