+++ src/components/AppTile.tsx
@@ -6,8 +6,8 @@ export function AppTile({
 	name,
 	index,
 	selected,
-	scrollOffset = 0,
 	onTouchEnd,
+	scrollOffset = 0,
 }: {
 	icon: ArrayBuffer | undefined;
 	name: string;
@@ -21,7 +21,7 @@ export function AppTile({
 	const width = root.ctx.canvas.width / perRow;
 	const height = width;
 	const x = (index % perRow) * width;
-	const y = Math.floor(index / perRow) * height - scrollOffset;
+	const y = Math.floor(index / perRow) * height - (scrollOffset * height) / 2;
 	const iconSize = width * 0.75;
 	return (
 		<Group width={width} height={height} x={x} y={y} onTouchEnd={onTouchEnd}>
