+++ src/components/FilePicker.tsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useState } from 'react';
 import { Group, Rect, Text, useRoot } from 'react-tela';
 import { useGamepadButton, useDirection } from '../hooks/use-gamepad';
 import { isDirectory } from '../util';
+import { Scrollbar } from './Scrollbar';
 
 interface Entry {
 	name: string;
@@ -195,44 +196,6 @@ export function FilePicker({ onSelect, onClose }: FilePickerProps) {
 				/>
 			</Group>
 		</>
-	);
-}
-
-interface ScrollbarProps {
-	height: number;
-	x: number;
-	numEntries: number;
-	itemsPerPage: number;
-	scrollOffset: number;
-	padding?: number;
-}
-
-function Scrollbar({
-	height,
-	x,
-	numEntries,
-	itemsPerPage,
-	scrollOffset,
-	padding = 8,
-}: ScrollbarProps) {
-	if (numEntries <= itemsPerPage) return null;
-
-	// Calculate scrollbar dimensions
-	const scrollbarWidth = 8;
-	const scrollbarHeight = Math.floor(
-		Math.max((height * itemsPerPage) / numEntries, height * 0.1),
-	);
-	const scrollbarY = Math.round(
-		(height - scrollbarHeight) * (scrollOffset / (numEntries - itemsPerPage)),
-	);
-	return (
-		<Rect
-			width={scrollbarWidth}
-			height={scrollbarHeight}
-			x={x - scrollbarWidth - padding}
-			y={scrollbarY}
-			fill='rgba(255, 255, 255, 0.5)'
-		/>
 	);
 }
